crc: move all crc APIs into one header
Move all CRC APIs into one single header.
